refactor: move BatchOperationExecutionScheduler a `scheduler` sub-package
