+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/batchoperation/scheduler/BatchOperationExecutionScheduler.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.engine.processing.batchoperation;
+package io.camunda.zeebe.engine.processing.batchoperation.scheduler;
 
 import com.google.common.base.Strings;
 import com.google.common.math.IntMath;
